Ignore the config file
The Discogs API requires the use of a 'key' and 'secret' (passphrase) for
app authorisation. Since these MUST not be shared publically, they are
instead made available to the code via environmental variables, which are
generated from the config.sh script (can later be adapted to e.g. Heroku).
Since this cannot be incorporated into source code, alternatives must be
provided for portablility (i.e. upload collection or personal auth).