Implement prototype Shelf class to store Records
Instead a list of Records with associated functions, a Shelf object
provides a good way to store and access Record objects. Shelf should
be initialisable with a variety of collection types (dataframe etc) but
is currently only by `collection_folder`. Storage of records in dict
is not yet for any purpose, and access has not been implemented in
discogs_jockey module.