Add `interactor` module with skeleton `Interactor` ABC
Interaction between the user and the game (i.e. user input, display
of game information, request for discogs details etc.) should be
handled seperately from the game code (and discogs API code).
Other modules can be written to rely on the existence of certain
functions for these purposes, hence the implementation of an
`Interactor` Abstract Base Class to enforce these methods,
allowing various different forms of display interface to be used
(e.g. Terminal, GUI). The ABC will be written in tandem with the
game, which is being rewritten.